pyproject.toml
[project]
name = "ai-saas-starter-kit"
version = "0.1.0"
description = "Enterprise AI Solutions Template"
requires-python = ">=3.11"

dependencies = [
    "django>=5.0",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "django-redis>=5.4",
    "celery>=5.3",
    "django-htmx>=1.17",
    "django-environ>=0.11.2",
    "httpx>=0.27",
    "whitenoise>=6.6",
    "gunicorn>=21.2.0",
    # Add flower for Celery monitoring
    "flower>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest-django>=4.5",
    "ipython>=8.12",
]
